These two tooltips were moved elsewhere in the new urlbar so ...
... so only the tooltip for the dropmarker should have been added back.
These other two do not appear to do anything; they were removed
inbetween firefox-77 and firefox-78, along with the desirable
dropmarker tooltip "show history". This patchset adds back the
dropmarker, as well as the dropmarker tooltip, but does not add back
the other removed stuff which was removed in:

https://github.com/mozilla/gecko-dev/commit/92345feaaa2602d289b751b1fafdeb3459cdbc37
https://phabricator.services.mozilla.com/D72467

So in short, this patchset did not need to add the other tooltips back,
as whatever had those two tooltips was reworked into something else.

I believe the reason I accidentally added them back was because there
was a conflict when applying the 77-patches for 78, and the conflict
was due to these lines being right above the dropmarker show history
line, and I solved the conflict by leaving them in rather than removing
them (as upstream 77->78 did).